Linux validation. Add modifications in CMAKE for OpenCL library, and validation for GCC and CLANG using Linux.

+++ src/geometries/GGEMSSolidBox.cc
@@ -118,9 +118,9 @@ GGEMSSolidBox::GGEMSSolidBox(GGsize const& virtual_element_number_x, GGsize cons
 
   if (opengl_manager.IsOpenGLActivated()) {
     opengl_solid_ = new GGEMSOpenGLParaGrid(
-      static_cast<GLint>(virtual_element_number_z),
-      static_cast<GLint>(virtual_element_number_y),
-      static_cast<GLint>(virtual_element_number_x),
+      virtual_element_number_z,
+      virtual_element_number_y,
+      virtual_element_number_x,
       box_size_z / static_cast<GLfloat>(virtual_element_number_z),
       box_size_y / static_cast<GLfloat>(virtual_element_number_y),
       box_size_x / static_cast<GLfloat>(virtual_element_number_x),
